+++ backend/packages/gitlab/CHANGELOG.md
@@ -1,5 +1,22 @@
 # @cat-factory/gitlab
 
+## 0.15.37
+
+### Patch Changes
+
+- Updated dependencies [e7867db]
+- Updated dependencies [00c4d94]
+  - @cat-factory/contracts@0.242.0
+  - @cat-factory/kernel@0.241.0
+
+## 0.15.36
+
+### Patch Changes
+
+- Updated dependencies [c5a1a16]
+  - @cat-factory/contracts@0.241.0
+  - @cat-factory/kernel@0.240.0
+
 ## 0.15.35
 
 ### Patch Changes