+++ backend/packages/provider-s3/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @cat-factory/provider-s3
 
+## 0.2.303
+
+### Patch Changes
+
+- Updated dependencies [e7867db]
+  - @cat-factory/kernel@0.241.0
+
+## 0.2.302
+
+### Patch Changes
+
+- Updated dependencies [c5a1a16]
+  - @cat-factory/kernel@0.240.0
+
 ## 0.2.301
 
 ### Patch Changes